Null check in AND operator. If left is false both sides of AND are executed.
Null check in AND operator. If left is false both sides of AND are executed.

Null check in AND operator. If left is false both sides of AND are executed.

Where x is and empty List of strings.
var x = new List();
"{{ if x && x.size > 0 }}{{ end }}";

Results into: ScriptRuntimeException, ": error : Object x is null. Cannot access member: orderitems.size".

https://github.com/lunet-io/scriban/issues/191